Fix integer overflow due to wrong type for field `connection.asn`
Java number types are signed. As a consequence, storing an unsigned int
value requires a long.

+++ src/main/java/co/ipregistry/api/client/model/Connection.java
@@ -27,7 +27,7 @@ import lombok.*;
 public class Connection {
 
     @JsonProperty("asn")
-    private Integer asn;
+    private Long asn;
 
     @JsonProperty("domain")
     private String domain;
